Tweet collection up to 42400 users now
A few more keys may be needed to catch all 46000 or so, but it's possible to get rid of the users with little or no followers or the celebrities who mostly provide retweets for our collection. The first few processes might throw 401 or 420 errors, but it smoothes out after a few seconds and the collection starts.

All tweets are now collected in the Data directory where each batch of 400 users has its own batch file where all the tweets are sent during streaming